made changes according to review comments and moved utility test interfaces into util package
